Round scale to nearest 26.6 fixed point.
With this change, the computation of the scale factor becomes
identical across the freetype and truetype packages, removing
deviations in the font metrics that are derived from scale.

The rounding computation is newer; it was introduced when the
truetype.Face type was added in commit 6deea2414309d03c665a7b.

Fixes #85.